finishing section 11

+++ cards/src/main/java/com/udemy/cards/repository/CardsRepository.java
@@ -7,7 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CardsRepository extends JpaRepository<Cards, Integer> {
+public interface CardsRepository extends JpaRepository<Cards, Long> {
+
     Optional<Cards> findByMobileNumber(String mobileNumber);
+
     Optional<Cards> findByCardNumber(String cardNumber);
+
 }
